application/repository/video_repository_test.go
package repository_test

import (
	"encoder-service/application/repository"
	"encoder-service/domain"
	"encoder-service/framework/database"
	"testing"
	"time"

	uuid "github.com/satori/go.uuid"
	"github.com/stretchr/testify/require"
)

func TestVideoRepositoryDbInsert(t *testing.T) {
	db := database.NewDBTest()
	defer db.Close()

	video := domain.NewVideo()
	video.ID = uuid.NewV4().String()
	video.FilePath = "path"
	video.CreatedAt = time.Now()

	repo := repository.VideoRepository{DB: db}
	repo.Insert(video)

	v,err := repo.Find(video.ID)

	require.NotEmpty(t, v.ID)
	require.Nil(t, err)
	require.Equal(t,v.ID, video.ID)
}